javadoc work

+++ bundles/org.eclipse.equinox.p2.engine/src/org/eclipse/equinox/internal/p2/engine/ProfileEvent.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007, 2008 IBM Corporation and others.
+ * Copyright (c) 2007, 2010 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -10,9 +10,8 @@
  *******************************************************************************/
 package org.eclipse.equinox.internal.p2.engine;
 
-import org.eclipse.equinox.p2.engine.IProfileEvent;
-
 import java.util.EventObject;
+import org.eclipse.equinox.p2.engine.IProfileEvent;
 
 /**
  * @noextend This class is not intended to be subclassed by clients.
@@ -21,9 +20,9 @@ import java.util.EventObject;
 public class ProfileEvent extends EventObject implements IProfileEvent {
 	private static final long serialVersionUID = 3082402920617281765L;
 
-	private byte reason;
+	private int reason;
 
-	public ProfileEvent(String profileId, byte reason) {
+	public ProfileEvent(String profileId, int reason) {
 		super(profileId);
 		this.reason = reason;
 	}
@@ -31,7 +30,7 @@ public class ProfileEvent extends EventObject implements IProfileEvent {
 	/* (non-Javadoc)
 	 * @see org.eclipse.equinox.p2.engine.IProfileEvent#getReason()
 	 */
-	public byte getReason() {
+	public int getReason() {
 		return reason;
 	}
